pyproject.toml
[project]
name = "mimir"
version = "0.1.0"
description = "AI Agent for product knowledge built on Norns."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "norns-sdk",
    "discord.py>=2.3",
    "PyGithub>=2.1",
    "google-api-python-client>=2.100",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.2",
]

[project.scripts]
mimir-worker = "mimir.worker:main"
mimir-discord = "mimir.discord_bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
norns-sdk = { path = "../norns-sdk-python", editable = true }
